fix(tabs): cycle issue
